+++ R/plotBoxMetric.R
@@ -4,8 +4,9 @@
 #' depending of bins.
 #' If the \code{bin} is absent from \code{gr}, a single boxplot is drawn.
 #'
-#' @param gr a GRanges input
-#' @param metric name of the column in \code{gr} metadata containing scores.
+#' @param rse a RangedSummarizedExperiment input. Aletrnatively: can be a
+#' GRanges object (for backward compatibility).
+#' @param metric name of the column in \code{rse} metadata containing scores.
 #' @param title title of the plot.
 #' @param trans_func A function to transform value of x before ploting.
 #' Useful to apply log10 transformation
@@ -29,26 +30,38 @@
 #'        title = "Metric"
 #'   )
 plotBoxMetric <- function(
-    gr,
+    rse,
     metric = "expr", title = "Metric",
     trans_func = function(x) x,
     ylim = NULL, ylab = "metric",
     palette = colorRampPalette(c("magenta", "black", "green"))
 ){
+    if (methods::is(rse, "RangedSummarizedExperiment")) {
+        gr <- SummarizedExperiment::rowRanges(rse)
+    } else  {
+        gr <- rse
+    }
+
     if (!is.null(gr$bin)) {
+        bin <- gr$bin
         graphics::boxplot(
             lapply(
-                stats::setNames(levels(factor(gr$bin)), levels(factor(gr$bin))),
-                function(i) trans_func(mcols(gr)[gr$bin == i, metric])
+                stats::setNames(
+                    levels(factor(bin)),
+                    levels(factor(bin))
+                ),
+                function(i) trans_func(
+                    S4Vectors::mcols(gr)[bin == i, metric]
+                )
             ),
             ylab = ylab, pch = 19, ylim = ylim,
-            col = palette(length(unique(gr$bin))), axes = FALSE
+            col = palette(length(unique(bin))), axes = FALSE
         )
-        axis(1, at = seq_along(levels(factor(gr$bin))),
-             labels = levels(factor(gr$bin)))
+        axis(1, at = seq_along(levels(factor(bin))),
+             labels = levels(factor(bin)))
     } else {
         graphics::boxplot(
-            trans_func(mcols(gr)[[metric]]),
+            trans_func(S4Vectors::mcols(gr)[[metric]]),
             ylab = ylab, pch = 19, ylim = ylim,
             col = palette(1), axes = FALSE
         )