+++ tests/testthat/test_plotMetric.R
@@ -1,0 +1,7 @@
+data("stackepi")
+test1 <- function() plotMetric(SummarizedExperiment::rowRanges(stackepi)$exp)
+
+test_that(
+    "default parameters", {
+    vdiffr::expect_doppelganger("plotMetric default", test1)
+})